+++ samplesheets/tests/test_permissions.py
@@ -44,7 +44,7 @@ class TestSampleSheetsPermissions(
         )
         self.study = self.investigation.studies.first()
         self.assay = self.study.assays.first()
-        self.ticket = self._make_irodsaccessticket(
+        self.ticket = self._make_ticket(
             project=self.project,
             path='/some/path',
             study=self.study,